// src/types/zotero-base-fields.ts
export const ZOTERO_FIELDS = [
    "DOI",
    "ISBN",
    "ISSN",
    "PMCID",
    "PMID",
    "abstractNote",
    "accessDate",
    "applicationNumber",
    "archive",
    "archiveID",
    "archiveLocation",
    "artworkMedium",
    "artworkSize",
    "assignee",
    "audioFileType",
    "audioRecordingFormat",
    "authority",
    "billNumber",
    "blogTitle",
    "bookTitle",
    "callNumber",
    "caseName",
    "citationKey",
    "code",
    "codeNumber",
    "codePages",
    "codeVolume",
    "committee",
    "company",
    "conferenceName",
    "country",
    "court",
    "date",
    "dateDecided",
    "dateEnacted",
    "dictionaryTitle",
    "distributor",
    "docketNumber",
    "documentNumber",
    "edition",
    "encyclopediaTitle",
    "episodeNumber",
    "eventPlace",
    "extra",
    "filingDate",
    "firstPage",
    "format",
    "forumTitle",
    "genre",
    "history",
    "identifier",
    "institution",
    "interviewMedium",
    "issue",
    "issueDate",
    "issuingAuthority",
    "journalAbbreviation",
    "label",
    "language",
    "legalStatus",
    "legislativeBody",
    "letterType",
    "libraryCatalog",
    "manuscriptType",
    "mapType",
    "medium",
    "meetingName",
    "nameOfAct",
    "network",
    "numPages",
    "number",
    "numberOfVolumes",
    "organization",
    "originalDate",
    "originalPlace",
    "originalPublisher",
    "pages",
    "partNumber",
    "partTitle",
    "patentNumber",
    "place",
    "postType",
    "presentationType",
    "priorityDate",
    "priorityNumbers",
    "proceedingsTitle",
    "programTitle",
    "programmingLanguage",
    "publicLawNumber",
    "publicationTitle",
    "publisher",
    "references",
    "reportNumber",
    "reportType",
    "reporter",
    "reporterVolume",
    "repository",
    "repositoryLocation",
    "rights",
    "runningTime",
    "scale",
    "section",
    "series",
    "seriesNumber",
    "seriesText",
    "seriesTitle",
    "session",
    "sessionTitle",
    "shortTitle",
    "status",
    "studio",
    "subject",
    "system",
    "thesisType",
    "title",
    "type",
    "university",
    "url",
    "versionNumber",
    "videoRecordingFormat",
    "volume",
    "websiteTitle",
    "websiteType",
] as const;